error check jail creation

+++ jackettinstall.sh
@@ -70,9 +70,13 @@ fi
 # Create Jail
 
 echo '{"pkgs":["nano","mono","curl","ca_root_nss"]}' > /tmp/pkg.json
-iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r ${RELEASE} ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
-
+if ! iocage create --name "${JAIL_NAME}" -p /tmp/pkg.json -r "${RELEASE}" ip4_addr="${INTERFACE}|${JAIL_IP}/24" defaultrouter="${DEFAULT_GW_IP}" boot="on" host_hostname="${JAIL_NAME}" vnet="${VNET}"
+then
+	echo "Failed to create jail"
+	exit 1
+fi
 rm /tmp/pkg.json
+
 iocage exec ${JAIL_NAME} sysrc jackett_enable="YES"
 iocage exec ${JAIL_NAME} service jackett start
